requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.29
pandas>=2.2.2
